chore: Split up asset generator pages
