+++ server/routes/generate-assets/badges.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+
+function GenerateBadges() {
+  return <div>generate badges</div>;
+}
+
+export default GenerateBadges;
