+++ server/routes/generate-assets/presentation.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+
+function GeneratePresentation() {
+  return <div>generate presentation</div>;
+}
+
+export default GeneratePresentation;
