+++ server/routes/generate-assets/schedule.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+
+function GenerateSchedule() {
+  return <div>generate schedule</div>;
+}
+
+export default GenerateSchedule;
