+++ server/routes/generate-assets/text.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+
+function GenerateText() {
+  return <div>generate text</div>;
+}
+
+export default GenerateText;
